#1760 Remove Weapon and RequestShootMessage and update  all prefabs to use Gun instead of Weapon.
