+++ UnityProject/Assets/Prefabs/Items/Weapons/Ballistics/Resources/C20rSMG_Ballistic.prefab
@@ -222,7 +222,7 @@ MonoBehaviour:
   Projectile: {fileID: 1932369758612010, guid: 4bcdc5dec532a42dba701d5162dab2ba, type: 3}
   ProjectilesFired: 1
   ProjectileVelocity: 10
-  WeaponType: 2
+  WeaponType: 1
 --- !u!1 &1992156404503320
 GameObject:
   m_ObjectHideFlags: 0
