+++ UnityProject/Assets/Prefabs/Items/Weapons/Ballistics/Resources/ERT_Gun_Ballistic.prefab
@@ -13,10 +13,10 @@ GameObject:
   - component: {fileID: 114703404969080964}
   - component: {fileID: 114346683018347606}
   - component: {fileID: 114906344895876820}
-  - component: {fileID: 114598204740365856}
   - component: {fileID: 114752816660759734}
   - component: {fileID: 114959013130737060}
   - component: {fileID: 114027104691969066}
+  - component: {fileID: 2251993831743313043}
   m_Layer: 10
   m_Name: ERT_Gun_Ballistic
   m_TagString: Untagged
@@ -138,30 +138,6 @@ MonoBehaviour:
   throwRange: 7
   hitSound: GenericHit
   attackVerb: []
---- !u!114 &114598204740365856
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1167865368234136}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 12ad712be1fefd043b45761b18e23798, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  AmmoType: 46x30mmtT
-  SpawnsCaseing: 1
-  SmartGun: 1
-  CurrentRecoilVariance: 0
-  FireCountDown: 0
-  FireingSound: ShootSMG
-  FireRate: 8
-  MaxRecoilVariance: 0.05
-  Projectile: {fileID: 1932369758612010, guid: 4bcdc5dec532a42dba701d5162dab2ba, type: 3}
-  ProjectilesFired: 1
-  ProjectileVelocity: 10
-  WeaponType: 2
 --- !u!114 &114752816660759734
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -223,6 +199,30 @@ MonoBehaviour:
       m_Calls: []
     m_TypeName: UnityEngine.Events.UnityEvent, UnityEngine.CoreModule, Version=0.0.0.0,
       Culture=neutral, PublicKeyToken=null
+--- !u!114 &2251993831743313043
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1167865368234136}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8343d0f678a24c949baac4bba12c2f59, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  AmmoType: 46x30mmtT
+  SpawnsCaseing: 1
+  SmartGun: 1
+  CurrentRecoilVariance: 0
+  FireCountDown: 0
+  FireingSound: ShootSMG
+  FireRate: 8
+  MaxRecoilVariance: 0.05
+  Projectile: {fileID: 1932369758612010, guid: 4bcdc5dec532a42dba701d5162dab2ba, type: 3}
+  ProjectilesFired: 1
+  ProjectileVelocity: 10
+  WeaponType: 1
 --- !u!1 &1205640307324434
 GameObject:
   m_ObjectHideFlags: 0
