+++ UnityProject/Assets/Prefabs/Items/Weapons/Ballistics/Resources/Handgun_Ballistic.prefab
@@ -92,10 +92,10 @@ GameObject:
   - component: {fileID: 114015923003440924}
   - component: {fileID: 114285597691290844}
   - component: {fileID: 114775921226298872}
-  - component: {fileID: 114078522730665838}
   - component: {fileID: 114586490647730340}
   - component: {fileID: 114437442048880876}
   - component: {fileID: 114956772244555426}
+  - component: {fileID: 4065799211730854973}
   m_Layer: 10
   m_Name: Handgun_Ballistic
   m_TagString: Untagged
@@ -217,30 +217,6 @@ MonoBehaviour:
   throwRange: 7
   hitSound: GenericHit
   attackVerb: []
---- !u!114 &114078522730665838
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1557459482202932}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 12ad712be1fefd043b45761b18e23798, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  AmmoType: 9mm
-  SpawnsCaseing: 1
-  SmartGun: 0
-  CurrentRecoilVariance: 0
-  FireCountDown: 0
-  FireingSound: ShootGun01
-  FireRate: 4
-  MaxRecoilVariance: 0
-  Projectile: {fileID: 1932369758612010, guid: 40ea4aea218be439eaf29f853612870c, type: 3}
-  ProjectilesFired: 1
-  ProjectileVelocity: 10
-  WeaponType: 1
 --- !u!114 &114586490647730340
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -302,3 +278,27 @@ MonoBehaviour:
       m_Calls: []
     m_TypeName: UnityEngine.Events.UnityEvent, UnityEngine.CoreModule, Version=0.0.0.0,
       Culture=neutral, PublicKeyToken=null
+--- !u!114 &4065799211730854973
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1557459482202932}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8343d0f678a24c949baac4bba12c2f59, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  AmmoType: 9mm
+  SpawnsCaseing: 1
+  SmartGun: 0
+  CurrentRecoilVariance: 0
+  FireCountDown: 0
+  FireingSound: ShootGun01
+  FireRate: 4
+  MaxRecoilVariance: 0
+  Projectile: {fileID: 1932369758612010, guid: 40ea4aea218be439eaf29f853612870c, type: 3}
+  ProjectilesFired: 1
+  ProjectileVelocity: 10
+  WeaponType: 0
