+++ UnityProject/Assets/Prefabs/Items/Weapons/Ballistics/Resources/Revolver_Ballistic.prefab
@@ -92,10 +92,10 @@ GameObject:
   - component: {fileID: 114173451881615334}
   - component: {fileID: 114184990726307632}
   - component: {fileID: 114461506418144158}
-  - component: {fileID: 114358469078938250}
   - component: {fileID: 114529648990615742}
   - component: {fileID: 114451834042202350}
   - component: {fileID: 114200900872640816}
+  - component: {fileID: 3377771713695131618}
   m_Layer: 10
   m_Name: Revolver_Ballistic
   m_TagString: Untagged
@@ -217,30 +217,6 @@ MonoBehaviour:
   throwRange: 7
   hitSound: GenericHit
   attackVerb: []
---- !u!114 &114358469078938250
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1863058284462694}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 12ad712be1fefd043b45761b18e23798, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  AmmoType: 357
-  SpawnsCaseing: 1
-  SmartGun: 0
-  CurrentRecoilVariance: 0
-  FireCountDown: 0
-  FireingSound: ShootGun01
-  FireRate: 3
-  MaxRecoilVariance: 0
-  Projectile: {fileID: 1932369758612010, guid: 008fa24b13fdb4e98971489ecee9218c, type: 3}
-  ProjectilesFired: 1
-  ProjectileVelocity: 10
-  WeaponType: 1
 --- !u!114 &114529648990615742
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -302,3 +278,27 @@ MonoBehaviour:
       m_Calls: []
     m_TypeName: UnityEngine.Events.UnityEvent, UnityEngine.CoreModule, Version=0.0.0.0,
       Culture=neutral, PublicKeyToken=null
+--- !u!114 &3377771713695131618
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1863058284462694}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8343d0f678a24c949baac4bba12c2f59, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  AmmoType: 357
+  SpawnsCaseing: 1
+  SmartGun: 0
+  CurrentRecoilVariance: 0
+  FireCountDown: 0
+  FireingSound: ShootGun01
+  FireRate: 3
+  MaxRecoilVariance: 0
+  Projectile: {fileID: 1932369758612010, guid: 008fa24b13fdb4e98971489ecee9218c, type: 3}
+  ProjectilesFired: 1
+  ProjectileVelocity: 10
+  WeaponType: 0
