+++ UnityProject/Assets/Prefabs/Items/Weapons/Ballistics/Resources/Shotgun_Ballistic.prefab
@@ -92,10 +92,10 @@ GameObject:
   - component: {fileID: 114284135461217894}
   - component: {fileID: 114308011406422678}
   - component: {fileID: 114677960272259078}
-  - component: {fileID: 114776078380092352}
   - component: {fileID: 114469553411307036}
   - component: {fileID: 114465474734708746}
   - component: {fileID: 114923350361433254}
+  - component: {fileID: 3241909802591281347}
   m_Layer: 10
   m_Name: Shotgun_Ballistic
   m_TagString: Untagged
@@ -217,30 +217,6 @@ MonoBehaviour:
   throwRange: 7
   hitSound: GenericHit
   attackVerb: []
---- !u!114 &114776078380092352
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1436741001650206}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 12ad712be1fefd043b45761b18e23798, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  AmmoType: Slug
-  SpawnsCaseing: 0
-  SmartGun: 0
-  CurrentRecoilVariance: 0
-  FireCountDown: 0
-  FireingSound: ShootLMG
-  FireRate: 1.5
-  MaxRecoilVariance: 0
-  Projectile: {fileID: 1932369758612010, guid: 1508ddac0ee0e46d5b6ec0f94ffb4179, type: 3}
-  ProjectilesFired: 1
-  ProjectileVelocity: 8
-  WeaponType: 1
 --- !u!114 &114469553411307036
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -302,3 +278,27 @@ MonoBehaviour:
       m_Calls: []
     m_TypeName: UnityEngine.Events.UnityEvent, UnityEngine.CoreModule, Version=0.0.0.0,
       Culture=neutral, PublicKeyToken=null
+--- !u!114 &3241909802591281347
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1436741001650206}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8343d0f678a24c949baac4bba12c2f59, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  AmmoType: Slug
+  SpawnsCaseing: 0
+  SmartGun: 0
+  CurrentRecoilVariance: 0
+  FireCountDown: 0
+  FireingSound: ShootLMG
+  FireRate: 1.5
+  MaxRecoilVariance: 0
+  Projectile: {fileID: 1932369758612010, guid: 1508ddac0ee0e46d5b6ec0f94ffb4179, type: 3}
+  ProjectilesFired: 1
+  ProjectileVelocity: 8
+  WeaponType: 0
