+++ UnityProject/Assets/Prefabs/Items/Weapons/Energy/Resources/Tazer.prefab
@@ -92,10 +92,10 @@ GameObject:
   - component: {fileID: 114958751264410972}
   - component: {fileID: 114154339735646076}
   - component: {fileID: 114374851331526312}
-  - component: {fileID: 114260592088361284}
   - component: {fileID: 114758040345321802}
   - component: {fileID: 114155970044334132}
   - component: {fileID: 114088013606370678}
+  - component: {fileID: 8735810588248712643}
   m_Layer: 10
   m_Name: Tazer
   m_TagString: Untagged
@@ -217,30 +217,6 @@ MonoBehaviour:
   throwRange: 7
   hitSound: GenericHit
   attackVerb: []
---- !u!114 &114260592088361284
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1392830670498528}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 12ad712be1fefd043b45761b18e23798, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  AmmoType: FusionCells
-  SpawnsCaseing: 0
-  SmartGun: 1
-  CurrentRecoilVariance: 0
-  FireCountDown: 0
-  FireingSound: ShootTazer
-  FireRate: 5
-  MaxRecoilVariance: 0
-  Projectile: {fileID: 1560634355228308, guid: 3a5f33fb4348d74429ba2cf1a1d4821d, type: 3}
-  ProjectilesFired: 1
-  ProjectileVelocity: 17
-  WeaponType: 1
 --- !u!114 &114758040345321802
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -302,3 +278,27 @@ MonoBehaviour:
       m_Calls: []
     m_TypeName: UnityEngine.Events.UnityEvent, UnityEngine.CoreModule, Version=0.0.0.0,
       Culture=neutral, PublicKeyToken=null
+--- !u!114 &8735810588248712643
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1392830670498528}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8343d0f678a24c949baac4bba12c2f59, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  AmmoType: FusionCells
+  SpawnsCaseing: 0
+  SmartGun: 1
+  CurrentRecoilVariance: 0
+  FireCountDown: 0
+  FireingSound: ShootTazer
+  FireRate: 5
+  MaxRecoilVariance: 0
+  Projectile: {fileID: 1560634355228308, guid: 3a5f33fb4348d74429ba2cf1a1d4821d, type: 3}
+  ProjectilesFired: 1
+  ProjectileVelocity: 17
+  WeaponType: 0
