--- UnityProject/Assets/Scripts/Messages/Client/RequestShootMessage.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 955e6c5fae8484bcfaff997da44cf04b
-timeCreated: 1529587316
-licenseType: Pro
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
