--- UnityProject/Assets/Scripts/Weapons/Weapon.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 12ad712be1fefd043b45761b18e23798
-timeCreated: 1494333900
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
